Remove `lastGamePaused` field from SharedMemoryService
The `lastGamePaused` field in the `SharedMemoryService` class
within the `R3E.API` namespace was removed. This field was
previously used to track the paused state of the game. Its
removal indicates that the paused state is no longer being
tracked in this manner, or the related functionality has been
refactored or deprecated.